Norminetted and refactored all of the codes including libft. 142/ 146 Solved.
