+++ Makefile
@@ -52,6 +52,9 @@ SRC_FILES := main.c \
                 builtins/builtin_unset.c \
 				pipex/execute_1.c \
 				pipex/execute_2.c \
+				pipex/execute_3.c \
+				pipex/execute_4.c \
+				pipex/execute_5.c \
 				pipex/execute_utils_1.c \
 				pipex/execute_utils_2.c \
 				pipex/get_command_array.c \
